added geolocation to device
