--- src/modules/geolocation/geolocation.module.ts
@@ -1,11 +1,0 @@
-import { Module } from '@nestjs/common';
-import { GeolocationController } from './geolocation.controller';
-import { GeolocationService } from './geolocation.service';
-
-@Module({
-  imports: [],
-  controllers: [GeolocationController],
-  providers: [GeolocationService],
-  exports: [GeolocationService],
-})
-export class GeolocationModule {}
